requirements.txt
textual>=0.86.0
httpx>=0.28.1
supabase>=2.28.2
python-dotenv>=1.2.2
textual-serve>=1.1.3
cryptography>=46.0.5
langgraph>=0.2.0
langchain-google-genai>=1.0.0
langchain-core>=0.3.0
biblia-cli==0.1.8
